Refactor constants ,colors , app theme and validator files move to utils, update imports
