+++ lib/core/apiManger/apiService.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:flowery/core/config/constants.dart';
+import 'package:flowery/core/utils/constants.dart';
 import 'package:retrofit/retrofit.dart';
 import 'package:json_annotation/json_annotation.dart';
