+++ lib/core/utils/application_theme.dart
@@ -1,4 +1,4 @@
-import 'package:flowery/core/config/colors.dart';
+import 'package:flowery/core/utils/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
